Add Board component and placeholder values
Placeholder values were used for the data passed to the Board component
to show client behaviour when starting a match. The structure of the
intended data is yet to be finalized (game logic) but what is added is
how the page would load and what it would display depending on the
response from the server.