Game init logic (#3)
* Add CLIENT_HOST for request origin validation

* Add WS endpoint for game

The 'cn' and 'url' variables are also attached to the socket to be used
in handling game actions (and spectating?)

* Add room-handling functions

Create a function to generate game data on game start. Store the data
to allow the client to determine whether or not the game (room) exists
before connecting to the WebSocket server

Upon visiting the game page, the client will query the database to
determine whether or not the game exists before connecting to the WS
server. Depending on the response, the client will either:
1. Display a "Page not found" notice
2. Connect to the WS server to get the current board state
3. Display the endgame board state taken from the database

Endgame board state will be stored so that the game state can be deleted
in the main 'gameStates' object (see /api/src/WSServer.ts)

* Add 2 hours lifespan to game

* Replace 'shortid' with 'nanoid'

'shortid' is now apparently deprecated; the notice was added to its
GitHub page just 4 months ago. Nanoid seems to be a much better
implementation too

* Add getGame and connectToGame functions for client

* Standardize variable/key name for room (should be 'id')

* Add Board component and placeholder values

Placeholder values were used for the data passed to the Board component
to show client behaviour when starting a match. The structure of the
intended data is yet to be finalized (game logic) but what is added is
how the page would load and what it would display depending on the
response from the server.