Add context providers for GameInfo and Socket
For 'GameInfo', if the game hasn't started yet, the pieces need the
'turn' value to determine whether or not they can move freely, that is,
for when setting up pieces, or if they can only move normally, that
is, when the game has already started.
The 'Panel' component on the other hand needs the 'GameInfo' for the
turn, player names, and countdowns after the game has started.

For 'Socket', in both cases (game start and game setup), the pieces need
access to the socket in order to be able to send data back to the
server. The use case for when the game has already started is obvious,
but during game setup, it's also necessary to transmit the current state
of the player's board in order to persist the player's setup on page
refresh.

To summarize:
Game setup:
Pieces - Turn (from GameInfo, will be undefined), Socket
Clock - Time (from GameInfo but doesn't need context)
Board - Board

Game start:
Pieces - Turn (from GameInfo), Socket
Panel - p1Info, p2Info, turn (all from GameInfo)
Board - Board