Add check for AFK players during setup
* Add SetupPanel component
* Add responses from WS Server for 'ready' and 'afk'
* Add 'deleteGame' function to remove room if AFK
* Add state changes for when socket is closed

WS Server closes the socket for everyone in the room if someone's AFK

* Fix some state-handling in /Board/index.jsx

Because SetupPanel was added, the time state should be passed down to
it. It should have been fixed beforehand though, because previously it
would have kept on rerendering the whole board since the state is
declared in the root component. It would be very inefficient considering
the component rerenders each .1 second due to the Countdown component.
The SetupPanel now gets the time from the server via a useContext hook