Add time-keeping implementation
* Fix 'Panel' component to prevent unnecessary rerendering

* Set 'idleTimeout' to 0 (disabling it) for WS server to deal with AFK

* Use 'performance.now()' instead of 'Date.now()' for accuracy

* Add 'TurnChangeReducer' to simplify state management in MatchClocks

* Add 'move' request and response to simulate change of turns
A 'move' case for the WS server response is added. A 'Board' component
is also added along with a new 'user' state in the parent component to
determine legality of move (also, to determine the CSS for the 'Players'
component to highlight who the client's name)

* Add 'lastMove' and 'lastClosed' key for time management
'lastClosed' is added to keep track of the time even while both
clients have the game closed. Its value is updated whenever the
socket is opened/closed. When reopening the page (reconnecting
the socket), the elapsed time from when the socket was last closed
to the current time is deducted to the remaining time of whomever's
turn it currently is. 'lastMove' is used the same way, except it
triggers when a move (in 'message') is made as opposed to when a
socket is closed.