Change setTimeout interval from 100 to 92.5
If it's set to 100, it's too slow compared to the time from the server.
Every 15 seconds, it lags behind an extra second compared to server time.
At 90, the client countdown becomes too fast. At 92.5, it gets delayed
by a second every 1 minute and 30 seconds. Until I can come up with a
more exact method to count the time without constantly badgering the
server (unless it's not that big of a deal?) this is sufficient because
if you're playing the game "legitimately", there's no way you'll spend
more than 30 seconds, or maybe even 20 seconds each turn.