Remove lastClosed key for time-keeping
It's unnecessary. 'lastMove' should just be updated each event (on
close, on start, on move) instead. If 'lastClosed' is to be included,
its value is supposed to be refreshed along with 'lastMove' each said
event anyway so it doesn't really make a difference except perhaps for
clarity. I should probably just change the key name or add a comment.
Probably.