feat: remake 'Panel' layout (#12)
*feat: display the winner of the match
Add 'winner' key to types. Add 'winner' column to tables. Fix response
to include the 'winner' column when querying the database. Add a check
for the winner in the server (for now, only condition is when time
reaches 0). Add a 'Winner' component for the frontend.

*feat: add CSS for responsive design
Use grids for the layout. Create media queries both for the grid and for
the size of the fonts

*refactor: fix components structure
Create a 'Board' and 'MatchClock' folder. Use a single 'Player' component
instead of separating 'MatchClocks' and 'Players'