feat: add checkIfLegal function for moves
Add a function to check if the move is legal. The actual verification
will still be done by the server since the values of the pieces would be
unknown, but the move can easily be checked beforehand. It lessens the
server hits by only allowing legal moves to be sent