fix: change name key to be optional
Board states should record the position and values of the player's
pieces. What will be published or sent to the player is the board state
that includes both his and the opponent's board. When merging both
boards, the player is (obviously) not allowed to know the values of the
pieces of the opponent. The name key is unnecessary for such cases (a
value of 1 is given)