fix: hide values of returned board on game start
The returned board should have its values scrubbed to prevent the player
from knowing the position of the pieces of the opponent. The server
publishes the board with its values removed while the receiver fixes the
values of its own pieces. Also, the 'ready' button in 'Setup.jsx' should
send the user's current board to the server.