fix: change structure of returned game data
Remove the 'board' parameter of 'getGameInfo', basically returning only
the 'metadata' (?) of the game, that is, anything not related to the
pieces. The published data by the WS server is also changed from putting
all the data into a single 'data' key, to separating the 'metadata' and
the board information. As for the case 'move' in the 'message' functoin
in 'game.ts', the board data contains the coordinates of the origin and
destination of the move. The handling of the returned data in the
client-side is adjusted accordingly with the changes made in the server