fix: prevent user from passing fake coordinates
checkIfLegal is already in the client. It'll be added to the server too just
in case the user managed to send fake coordinates like 'X1' or something
like 'A1' as the origin but 'C5' as the destination. The checks are
still done in the client as a way of preventing the user from repeatedly
sending bogus moves to the server (reducing unnecessary hits)