feat: add pause to time on move
By setting 'turn' to null, it'll make the countdown stop ticking. It'll
help with the disparity in time between the client and the server
caused by delay in message transfer (from client to server and vice
versa). There will still be a difference because of the delay from
client to server but at least the extra difference caused by the delay
from server to client is gone.