feat: add 'getCookie' function
The function will run in the homepage if the user is not logged in or if
the user doesn't have a previous cookie (given by the function).
'withAuthGSSP' is modified to look for the '=' character in order to
determine of the cookie is from logging in or is just from the
'getCookie' function. Normally, 'decrypting' the token inside the cookie
will just return the username. However, if it's the cookie from the
function, its output will instead be a 15-character string prefixed by
the characters 'c=', in order to differentiate it.