fix: prevent all Piece components from rerendering
Previously, on each move, every Piece component, even the ones not
involved in the "move", were rerendering. Instead of relying on a state
from the parent component (Pieces), a state 'dragging' was set in the
Piece component that would replace the behaviour of the previous
'draggable' state from Pieces.
The useContext hook was also removed in favor of simply passing down the
dispatch and the relevant state (winner) as props, removing DragContext
altogether