fix: make socket.send trigger properly
The dependency for useEffect originally was just the turn. This means
that if after the 'time' case was sent to the server but the player's
time in the server hasn't actually ran out yet, the game data will be
sent back but the clock won't update from it's previous 0 time remaining
because turn hasn't changed. Adding gameInfo.[playerNum].time solves
that problem

Also, with regards to the discrepancy, sometimes the time sent back is
below zero because of the delay in the arrival of the message from the
client to the server. The countdown function (that triggers socket.send)
should be checking for <= zero instead of exactly for zero