fix: add conditional for message type 'time'
When time goes to zero in the client, it triggers a socket.send even if
the room doesn't exist anymore. Normally, the server is supposed to
close the socket connection and stop the function call. A check for
message type 'time' is added to allow the connection to persist (so the
display data postgame isn't erased)