fix: return null if key is undefined
There's some bug that causes the pieces on previously occupied
coordinates to persist after setup, showing up as the opponent's pieces
(class 'unknown'). I'm not really sure if this fixes it
since even if the keys aren't deleted, it shouldn't be an 'unknown'
class since the 'name' prop would be undefined if the problem is what
this commit is trying to fix. Huh.