fix: make element stick to cursor on mouse move
fixes #51

onDrag is unreliable so I switched to this method. Using onMouseMove
works better but if the cursor moves away from the dragged Piece too
fast, the mouseMove event stops. It needs a bigger container to work
properly, hence the mouseMove event handler being placed in DropTarget.
However, the event needs to recognize the dragged Piece, which is why
the setState function of the piece is passed to the dispatch, so the
update of the coordinates of the Piece (in order to make it stick to
the cursor) is done via setState.

I hope the referencing to the GitHub issue worked.