fix: user can't connect to the lobby immediately
Going to a lobby created by someone else doesn't immediately connect you
to the WebSocket if you don't have a cookie. The useCookie hook doesn't
work; it doesn't attach the cookie immediately making you join the lobby
on page visit even if it's the first time you visited the website (for
example, if a friend sent them the link so they can play against each
other).

A button is added for such cases. This way, they can just "force" the
connection. In any case, the useEffect hook is still left behind for
users who DO have a cookie before going to the lobby.