feat: add handler for status indicator querying
Add 'connections' to IRoom interface

Remove deletion of the game state on game end. It crashes the server
because everything is set to undefined after deletion, encountering
errors when querying deep properties.
Just leave the cleanup of game states to the cronjob

Add case 'ping' for status indicator querying for in socket onmessage handler