fix: rework lobbies
BREAKING CHANGE: changed tables in the database

Leave cleanup to cronjob instead of deleting on close. It allows
players to rejoin the lobby instead of immediately deleting it once the
player closes the tab.

In order to make the cleanup work, add an expiry. This way, the lobby
can persist for a short while. When querying the lobby, just return the
lobbies that haven't expired yet. On socket close, just remove the
socket from the connections array.

However, if someone else joins the lobby to start the match, delete the
lobby both in memory and in the database since it's not needed anymore