fix: remove cookieValue props
It's causing issues because the cookie doesn't get passed down to the
components even after using 'useCookie'. With the useButton hook, the
button is disabled before the component mounts so the check with the
'auth' parameters is now unnecessary (although it was also unnecessary
before since the request will just get declined because of the
'verifyToken' middleware of the createLobby endpoint).

There's still a brief amount of time where it causes errors (button text
to 'Please try again') because there's a delay from page load till the
useCookie hook takes effect, but it's less than a second. If the user
who visited the page for the first time (that is, no cookie yet) clicked
on the CreatePrivateLobby button after a second from page load, there
aren't any issues.