fix: change logic in 'afk' case
Do not delete the game state in memory. It crashes the server because of
the status indicator repeatedly querying it. The cleanup should be left
to the cronjob. Maybe add an extra parameter for the storeGame function
to change its expiry to a minute from storing the game instead of
waiting the original full 2 hours?