fix: force refresh if AFK
With how it currently works, the cronjob deletes the expired games both
in memory and in the database. The problem is that as long as the user
is in the page, the status indicator will keep on doing its job,
querying the WS '/ingameCount' endpoint. While the 'AFK' case doesn't
delete the game state in memory immediately anymore, if the player stays
on the page, it'll just crash the server again. Forcing a refresh seems
to be the way to go for now.

There definitely should be something added to the status indicator to
prevent it from querying if the game has ended already.