feat: add connectionHandler function
The HTTP status code set during (decline of) the handshake can't be
accessed by the WebSocket in the client. It basically requires me
to accept the handshake and then add the logic that would've
normally been in the handshake into the 'socket.onopen' function.
That specific logic is basically this `connectionHandler` function.

It closes the socket via `socket.end(code, reason)` if it doesn't
meet the requirements (subject to change). The code sent via
`socket.end` is received by the `onclose` handler in the client to
determine what to do depending on what error happened.