fix: rework sockets
Remove the previous retry implementation since it was buggy asda asd asd
Remove the previous retry implementation since it was buggy

Properly close the socket on useEffect cleanup (specify if it's open,
checked with `readyState` instead of relying on `state.ongoing`)

Fix checking of connections for the status indicator. Previously, it
just refreshes the state by rerunning useEffect with the stringified
dependency. It's sometimes unreliable (difficult to recreate so I wasn't
sure what exactly goes wrong, I think it's because the socket isn't
connecting on time)

Instead, a boolean is placed in the dependency array. It'll
automatically trigger useEffect whenever the incoming connections list
and the saved connections list have a different length because of the
`!connections.retry` part. Hopefully it works properly now...

A new state `text` is added to display the current connection state.
It's initially set to 'Connecting to the game...' but it can be changed
to either 'Game not found' or 'The game is currently ongoing' depending
on the error reason sent by the server.