feat: allow spectators
Change connectionhandler to allow spectators, denying the connection
only if the game does not exist or if the game is still in its setup
phase. Previously, it would deny the connection if you weren't
included in the list of players for that game.

Currently, if it's a bot game, you can already see the pieces of the
player. It's pretty useless for spectating games between two players
because I've yet to add the feature to allow you to share your
perspective; it's pretty much only useful for bot games right now. If
you 'spectate' a game (that is, go to a link with a game between two
players), you'll end up seeing 'unknown' pieces. This is the intended
behaviour. There will soon be an option for the player to allow
specific users to see his pieces. This commit basically sets that up.