fix: fix game handler
Use handshakeHandler, SendSocketMessage, and declareWinner to refactor
code.

Change cleanBoards usage because it no longer returns a value.

Add extra response for bot games. If the player is in a bot game, after
sending a move, the bot's move in response is already generated. It will
publish it immediately after the player's move. In the client, there's a
delay in fixing the board if the move is a bot move so that the bot move
doesn't instantly gets displayed. It's mostly for UX when playing
against bots. The delay is set in the client because it's easier to
manage (especially since I already have a custom hook specifically for
that purpose) than adding the delay in the server