fix: change socket.cn value
Use the IP address of the user if the user has no cookie. This is
necessary for users who directly visit a game link instead of going
through the index page to get a cookie. The useCookie hook can be
included but it wouldn't work anyway because the socket connects to the
WS server immediately, before the useCookie hook can attach the cookie
to the browser.

It's not that big of a deal because what will just happen
is that if you gave your PIN to someone else, if he tries to spectate
your game, everyone with the same address will be able to see your
board. I should probably fix it regardless, but maybe next time...